requirements.txt
numpy >= 1.17.0
Pillow >= 6.2.0
